Revert "test(vue-query): use vitest typecheck correctly with *.test-d.ts (#6988)"
This reverts commit 9da12e2b97f66c2204ec6b9a9a48437bf8c746b1.
